requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.14.3
boto3==1.42.16
botocore==1.42.16
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
jmespath==1.0.1
lxml==5.4.0
packaging==25.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redshift-connector==2.1.10
requests==2.32.5
s3transfer==0.16.0
scramp==1.4.6
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
typing_extensions==4.15.0
urllib3==2.6.2
